+++ examples/formatting-js/examples/quote_headers_array.example.js
@@ -1,0 +1,18 @@
+const csv = require('@fast-csv/format');
+
+const csvStream = csv.format({ headers: ['header1', 'header2'], quoteHeaders: [false, true] });
+
+csvStream.pipe(process.stdout).on('end', process.exit);
+
+csvStream.write({ header1: 'value1a', header2: 'value2a' });
+csvStream.write({ header1: 'value1a', header2: 'value2a' });
+csvStream.write({ header1: 'value1a', header2: 'value2a' });
+csvStream.write({ header1: 'value1a', header2: 'value2a' });
+csvStream.end();
+
+// Output:
+// header1,"header2"
+// value1a,value2a
+// value1a,value2a
+// value1a,value2a
+// value1a,value2a