+++ jest.config.js
@@ -4,8 +4,6 @@ module.exports = {
     // Replace `ts-jest` with the preset you want to use
     // from the above list
     preset: 'ts-jest',
-    collectCoverageFrom: [
-        "src/**/*.ts"
-    ],
-    testMatch: [ "**/__tests__/**/*.spec.ts"],
+    collectCoverageFrom: ['packages/**/*.ts', '!**/__tests__/**', '!**/build/**', '!**/node_modules/**'],
+    testMatch: ['**/__tests__/**/*.spec.ts'],
 };