+++ packages/parse/src/parser/column/ColumnParser.ts
@@ -1,0 +1,28 @@
+import { ParserOptions } from '../../ParserOptions';
+import { NonQuotedColumnParser } from './NonQuotedColumnParser';
+import { QuotedColumnParser } from './QuotedColumnParser';
+import { Scanner } from '../Scanner';
+import { Token } from '../Token';
+
+export class ColumnParser {
+    private readonly parserOptions: ParserOptions;
+
+    public readonly nonQuotedColumnParser: NonQuotedColumnParser;
+
+    public readonly quotedColumnParser: QuotedColumnParser;
+
+    public constructor(parserOptions: ParserOptions) {
+        this.parserOptions = parserOptions;
+        this.quotedColumnParser = new QuotedColumnParser(parserOptions);
+        this.nonQuotedColumnParser = new NonQuotedColumnParser(parserOptions);
+    }
+
+    public parse(scanner: Scanner): string | null {
+        const { nextNonSpaceToken } = scanner;
+        if (nextNonSpaceToken !== null && Token.isTokenQuote(nextNonSpaceToken, this.parserOptions)) {
+            scanner.advanceToToken(nextNonSpaceToken);
+            return this.quotedColumnParser.parse(scanner);
+        }
+        return this.nonQuotedColumnParser.parse(scanner);
+    }
+}