+++ src/parser/parser/index.ts
@@ -1,5 +1,0 @@
-export { default as Parser } from './Parser';
-export { default as RowParser } from './RowParser';
-export { Scanner } from './Scanner';
-export { Token, MaybeToken } from './Token';
-export { ColumnParser, NonQuotedColumnParser, QuotedColumnParser } from './column';